requirements.txt
fastapi
uvicorn
pydantic
requests
pytest
google-generativeai
pandas